requirements.txt
PyMySQL==1.1.1
requests==2.32.3
beautifulsoup4==4.13.5
playwright==1.55.0
pika==1.3.2
python-dotenv==1.1.1
cryptography==45.0.5
Django==5.2.5
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
